+++ ros_ws/launch/qlearning.launch
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <launch>
   <arg name="gui" default="true"/>
-  <arg name="realtime_simulation" default="true"/>
+  <arg name="realtime_simulation" default="false"/>
 
   <include file="$(find racer_world)/launch/racer_gazebo.launch">
       <arg name="world" value="racetrack_decorated"/>
@@ -10,7 +10,7 @@
       <arg name="verbose" value="false"/>
       <arg name="use_gpu" value="true"/>
       <arg name="extra_gazebo_args" value="--profile qlearning" unless="$(arg realtime_simulation)"/>
-      <arg name="laser_sample_count" value="32"/>
+      <arg name="laser_sample_count" value="8"/>
   </include>
 
   <include file="$(find vesc_sim)/launch/vesc_sim.launch"/>
@@ -25,7 +25,7 @@
 
   <node
     pkg="q_learning"
-    type="drive.py"
+    type="policyGradient.py"
     name="q_learning_training"
     output="screen" >
   </node>
