+++ ros_ws/src/autonomous/q_learning/scripts/parameters.py
@@ -16,12 +16,12 @@ TOPIC_GAZEBO_MODEL_STATE = "/gazebo/model_states"
 
 # General parameters
 
-ACTIONS = [(-0.9, 0.2), (0.9, 0.2),(0.5,0.2),(-0.5,0.2),(0, 0.5)]
+ACTIONS = [(-0.8, 0.1), (0.8, 0.1),(0.5,0.2),(-0.5,0.2),(0,0.2),(0, 0.4)]
 ACTION_COUNT = len(ACTIONS)
 
 # Only use some of the LIDAR measurements
 # When changing this value, also update laser_sample_count in qlearning.launch
-LASER_SAMPLE_COUNT = 32
+LASER_SAMPLE_COUNT = 8
 
 
 MODEL_FILENAME = os.path.join(RosPack().get_path("q_learning"), "model.to")
@@ -47,7 +47,46 @@ class NeuralQEstimator(nn.Module):
     def save(self):
         torch.save(self.state_dict(), MODEL_FILENAME)
 
+class Policy(nn.Module):
+    def __init__(self):
+        super(Policy, self).__init__()
+        self.state_space = LASER_SAMPLE_COUNT
+        self.action_space = ACTION_COUNT
+        
+        self.fc1 = nn.Linear(LASER_SAMPLE_COUNT, 32)
+        self.fc2 = nn.Linear(32, 18)
+        self.fc3 = nn.Linear(18, ACTION_COUNT)
 
+        self.gamma = 0.99
+
+        # Episode policy and reward history 
+        self.policy_history = Variable(torch.Tensor()) 
+        self.reward_episode = []
+
+        # Overall reward and loss history
+        self.reward_history = []
+        self.loss_history = []
+        
+
+    def forward(self, x):    
+        model = torch.nn.Sequential(
+            self.fc1,
+            nn.ReLU(),
+            self.fc2,
+            nn.ReLU(),
+            self.fc3,
+            nn.Softmax(dim=-1)
+        )
+        return model(x)
+
+    def load(self):
+        if os.path.isfile(MODEL_FILENAME):
+            self.load_state_dict(torch.load(MODEL_FILENAME))
+            rospy.loginfo("Model parameters loaded.")
+
+    def save(self):
+        torch.save(self.state_dict(), MODEL_FILENAME)
+        
 # Training parameters
 
 # Start by loading previously trained parameters.
@@ -62,7 +101,7 @@ MAX_EPISODE_LENGTH = 5000
 MEMORY_SIZE = 5000
 
 BATCH_SIZE = 128
-LEARNING_RATE = 0.0001
+LEARNING_RATE = 0.001
 
 # Probability to select a random episode starts at EPS_START
 # and reaches EPS_END once EPS_DECAY episodes are completed.
