+++ ros_ws/src/autonomous/q_learning/scripts/policyGradient.py
@@ -1,227 +1,228 @@
 #!/usr/bin/env python
+# coding: utf-8
 
+from qlearning import QLearningNode, device
+import random
+import math
+import time
+from collections import deque
+from std_msgs.msg import Empty
+from parameters import *
 import numpy as np
 
-import torch
-import torch.nn as nn
-import torch.optim as optim
-import torch.nn.functional as F
 from torch.autograd import Variable
 from torch.distributions import Categorical
 
-import math
-import random
-import rospy
-import car
+import torch
+import torch.optim as optim
+import torch.nn.functional as F
 
-rospy.init_node('learn', anonymous=True)
+from simulation_tools.reset_car import Point
+import simulation_tools.reset_car as reset_car
+from simulation_tools.track import track
 
+from gazebo_msgs.msg import ModelStates
 
-ACTIONS = [(-0.4, 0.1),(0.4, 0.1)]
-#ACTION_COUNT = len(ACTIONS)
+BATCH_INDICES = torch.arange(0, BATCH_SIZE, device=device, dtype=torch.long)
 
-# angle space are 18 possible steering directions from -1 to 1
-#angle_space = np.linspace(start=-1, stop=1, num=18, endpoint=True)
-#angle_space_len = len(angle_space)
-#angle_space = np.insert(angle_space, (angle_space_len/2), 0.0)
-
-# velocity space are the possible speed settings for the different steering directions
-# with higher speeds for smaller angles
-#velocity_space_half = np.concatenate((
-#        np.repeat(0.05, 6),
-#        np.repeat(0.1, 2),
-#        np.repeat(0.15, 1)))
-#velocity_space = np.concatenate(
-#    (velocity_space_half,
-#    [2.0],
-#    np.flip(velocity_space_half)))
-   
-#ACTIONS = zip(angle_space, velocity_space)
-ACTION_COUNT = len(ACTIONS)
-
-#print(ACTION_COUNT)
-#print(ACTIONS)
-
-LASER_SAMPLE_COUNT = 32  # Only use some of the LIDAR measurements
-
-learning_rate = 0.01
-gamma = 0.99
-running_reward = 0
-done = False
-crash = False
-deadlock = 0
-
-class Policy(nn.Module):
+class QLearningTrainingNode(QLearningNode):
     def __init__(self):
-        super(Policy, self).__init__()
-        self.state_space = LASER_SAMPLE_COUNT
-        self.action_space = ACTION_COUNT
+        QLearningNode.__init__(self)
         
-        self.l1 = nn.Linear(self.state_space, 64, bias=False)
-        self.l2 = nn.Linear(64, self.action_space, bias=False)
+        self.policy1 = Policy()
+        self.optimizer1 = optim.Adam(self.policy1.parameters(), lr=LEARNING_RATE)
+
+        self.episode_count = 0
+        self.episode_length = 0
+        self.total_step_count = 0
+        self.optimization_step_count = 0
+        self.cumulative_reward = 0
+        self.is_terminal_step = False
+
+        self.net_output_debug_string = ""
+        self.episode_length_history = deque(maxlen=50)
+        self.cumulative_reward_history = deque(maxlen=50)
+
+        self.state = None
+        self.action = None
+        self.car_position = None
+        self.car_orientation = None
+
+        self.drive_forward = None
+        self.steps_with_wrong_orientation = 0
+
+        self.real_time_factor = 0
+        self.episode_start_time_real = time.time()
+        self.episode_start_time_sim = rospy.Time.now().to_sec()
+
+        if USE_EXISTING_PARAMETERS:
+            self.policy1.load()
+
+        self.optimizer = optim.Adam(self.policy.parameters(), lr=LEARNING_RATE)
+        self.memory = deque(maxlen=MEMORY_SIZE)
+        reset_car.register_service()
+
+        rospy.Subscriber(TOPIC_CRASH, Empty, self.on_crash)
+        rospy.Subscriber(TOPIC_GAZEBO_MODEL_STATE, ModelStates, self.on_model_state_callback)  # nopep8
+
+    def update_policy(self):
+        R = 0
+        rewards = []
+
+        # Discount future rewards back to the present using gamma
+        for r in self.policy1.reward_episode[::-1]:
+            R = r + self.policy1.gamma * R
+            rewards.insert(0,R)
+
+        # Scale rewards
+        rewards = torch.FloatTensor(rewards)
+        if rewards.numel() > 1:
+            rewards = (rewards - rewards.mean()) / (rewards.std() + np.finfo(np.float32).eps)
+
+        # Calculate loss
+        loss = (torch.sum(torch.mul(self.policy1.policy_history, Variable(rewards)).mul(-1), -1))
+
+        # Update network weights
+        if loss != 0:
+            self.optimizer1.zero_grad()
+            loss.backward()
+            self.optimizer1.step()
+
+            #Save episode history counters
+            self.policy1.loss_history.append(loss.item())
+            self.policy1.reward_history.append(np.sum(self.policy1.reward_episode))
         
-        self.gamma = gamma
-        
-        # Episode policy and reward history 
-        self.policy_history = Variable(torch.Tensor()) 
-        self.reward_episode = []
-        # Overall reward and loss history
-        self.reward_history = []
-        self.loss_history = []
+        # Reset policy history and rewards
+        self.policy1.policy_history = Variable(torch.Tensor())
+        self.policy1.reward_episode= []
 
-    def forward(self, x):    
-        model = torch.nn.Sequential(
-            self.l1,
-            nn.Dropout(p=0.6),
-            nn.ReLU(),
-            self.l2,
-            nn.Softmax(dim=-1)
-        )
-        return model(x)
+    def on_crash(self, _):
+        global done
+        if self.episode_length > 5:
+            self.is_terminal_step = True
+            done = True
 
-policy = Policy()
-optimizer = optim.Adam(policy.parameters(), lr=learning_rate)
+    def select_action(self, state):
+        #Select an action (0 or 1) by running policy model and choosing based on the probabilities in state
+        state = self.state.type(torch.FloatTensor)
+        state = self.policy1(Variable(state))
+        c = Categorical(state)
+        action = c.sample()
 
-def select_action(state):
-    #Select an action (0 or 1) by running policy model and choosing based on the probabilities in state
-    state = state.type(torch.FloatTensor)
-    state = policy(Variable(state))
-    
-    c = Categorical(state)
-    
-    action = c.sample()
-    
-    # Add log probability of our chosen action to our history    
-    if policy.policy_history.dim() != 0:
-        policy.policy_history = torch.cat([policy.policy_history, c.log_prob(action).view(1)])
-    else:
-        policy.policy_history = (c.log_prob(action))
-    return action
-
-def update_policy():
-    R = 0
-    rewards = []
-    
-    # Discount future rewards back to the present using gamma
-    for r in policy.reward_episode[::-1]:
-        R = r + policy.gamma * R
-        rewards.insert(0,R)
-        
-    # Scale rewards
-    rewards = torch.FloatTensor(rewards)
-    
-    ##########DEBUG
-    #rospy.loginfo("BEFORE rewards: " + str(rewards))
-    rewards1 = rewards - rewards.mean()
-    #rospy.loginfo("rewards1: " + str(rewards1))
-    rewards2 = rewards.std() + np.finfo(np.float32).eps
-    rospy.loginfo("rewards std: " + str(rewards.std()))
-    rospy.loginfo("rewards finfo: " + str(np.finfo(np.float32).eps))
-    
-    if rewards.numel() > 1:
-        rewards = (rewards - rewards.mean()) / (rewards.std() + np.finfo(np.float32).eps)
-
-    #print('REWARDS: {}'.format(rewards))
-    
-    #rospy.loginfo("AFTER rewards: " + str(rewards))
-    
-    # Calculate loss
-    loss = (torch.sum(torch.mul(policy.policy_history, Variable(rewards)).mul(-1), -1))
-
-    print('LOSS: {}'.format(loss))
-    
-    # Update network weights
-    optimizer.zero_grad()
-    loss.backward()
-    optimizer.step()
-    
-    #Save and intialize episode history counters
-    policy.loss_history.append(loss.item())
-    policy.reward_history.append(np.sum(policy.reward_episode))
-    policy.policy_history = Variable(torch.Tensor())
-    policy.reward_episode= []
-
-def perform_action(action_index):
-    if action_index < 0 or action_index >= len(ACTIONS):
-        raise Exception("Invalid action: " + str(action_index))
-    car.drive(*ACTIONS[action_index])
-
-def on_crash():
-    global crash
-    rospy.loginfo("testing...")
-    crash = True
-
-## TODO REWARDS besser skalieren
-## Maybe zu viele actions sodass zu kompliziert
-
-def main(episodes, device):
-    global done, crash, deadlock 
-    running_reward = 1
-    average_reward_per_episode = 0
-    for episode in range(episodes):
-
-        # Spawns the car the coordinates (x=0, y= -0.5) with an orientation of math.pi + (0 if random.random() > 0.5 else math.pi)
-        car.reset((0, -0.5), math.pi + (0 if random.random() > 0.5 else math.pi))
-        state = car.get_scan(LASER_SAMPLE_COUNT, device)
-        
-        # If true, the current episode will end instantly
-            
-        for time in range(10000):
-            done = False  
-            crash = False
-            # Choose an action w.r.t the current state
-            action = select_action(state)
-            # Step through environment using chosen action
-            perform_action(action)
-
-            # Get the next state
-            state = car.get_scan(LASER_SAMPLE_COUNT, device)
-            # The faster the car is driving, the bigger the reward
-            running_reward = 10 * ACTIONS[action][1]
-            
-            # True when the car intersects with a wall 
-            if crash:
-                running_reward = -10000 * ACTIONS[action][1] * 2000/(time+1)
-                done = True
-                rospy.loginfo("crash")
-
-            # True if car is flipped 
-            car_orientation = car.get_car_pos_x()
-            if (car_orientation[1] < -0.2 and car_orientation[1] > -0.322):
-                running_reward = -3000
-                done = True
-
-            # End the episode if the car is driving against a wall for more than 4 steps
-            if running_reward < 0:
-                deadlock += 1
-                if deadlock >= 5:
-                    done = True
-                    deadlock = 0
-            else:
-                deadlock = 0
-            
-            # Save reward
-            policy.reward_episode.append(running_reward)
-            average_reward_per_episode += running_reward
-            # End the episode
-            if done:
-                break
-        
-        # Used to determine when the environment is solved.
-        #running_reward = (running_reward * 0.99) + (time * 0.01)
-        if(time == 0):
-            policy.reward_episode = []
-            policy.policy_history = Variable(torch.Tensor())
+        # Add log probability of our chosen action to our history
+        if self.policy1.policy_history.dim() != 0:
+            self.policy1.policy_history = torch.cat([self.policy1.policy_history, c.log_prob(action).view(1)])
         else:
-            update_policy()
+            self.policy1.policy_history = (c.log_prob(action))
+        return action
 
-        print('Episode {}\tLast length: {:5d}\tSum Reward: {}'.format(episode, time, average_reward_per_episode))
-        average_reward_per_episode = 0
+    def get_reward(self):
+        # The faster the car, the more reward it gets
+        scaleForSpeed = 0
+        if self.action < 2:
+            scaleForSpeed = 0.4
+        elif self.action < 5:
+            scaleForSpeed = 0.7
+        else:
+            scaleForSpeed = 1
 
-rospy.loginfo("Initializing Pytorch...")
-device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+        track_position = track.localize(self.car_position)
+        distance = abs(track_position.distance_to_center)
 
-car.register_crash_callback(on_crash)
-episodes = 1000
-main(episodes, device)
+        # The closer the car is to the middle of the race track, the more reward it gets
+        if distance < 0.2:
+            return 1 * scaleForSpeed
+        elif distance < 0.4:
+            return 0.7 * scaleForSpeed
+        else:
+            return -0.4 * scaleForSpeed
+        
+    def log_training_progress(self):
+        average_episode_length = sum(
+            self.episode_length_history) / len(self.episode_length_history)
+        average_cumulative_reward = sum(
+            self.cumulative_reward_history) / len(self.cumulative_reward_history)
+        rospy.loginfo("Episode " + str(self.episode_count) + ": "  # nopep8 \
+            + str(self.episode_length).rjust(3) + " steps (" + str(int(average_episode_length)).rjust(3) + " avg), "  # nopep8 \
+            + "return: " + "{0:.1f}".format(self.cumulative_reward).rjust(5)  # nopep8 \
+            + " (" + "{0:.1f}".format(average_cumulative_reward).rjust(5) + " avg), "  # nopep8 \
+            + ("memory: {0:d} / {1:d}, ".format(len(self.memory), MEMORY_SIZE) if len(self.memory) < MEMORY_SIZE else "")  # nopep8 \
+            + "time: {0:.1f}x, ".format(self.real_time_factor)  # nopep8 \
+            + "laser: {0:.1f} Hz".format(float(self.episode_length) / (rospy.Time.now().to_sec() - self.episode_start_time_sim))  # nopep8 \
+            )  # nopep8
 
+    def on_complete_episode(self):
+        self.update_policy()
+
+        self.episode_length_history.append(self.episode_length)
+        self.cumulative_reward_history.append(self.cumulative_reward)
+
+        real_time = time.time()
+        sim_time = rospy.Time.now().to_sec()
+        self.real_time_factor = (sim_time - self.episode_start_time_sim) / (real_time - self.episode_start_time_real)  # nopep8
+        self.log_training_progress()
+        self.episode_start_time_real = real_time
+        self.episode_start_time_sim = sim_time
+
+        self.episode_count += 1
+        self.episode_length = 0
+        self.cumulative_reward = 0
+
+        if self.episode_count % 50 == 0:
+            self.policy1.save()
+            rospy.loginfo("Model parameters saved.")
+
+    def check_car_orientation(self):
+        if self.car_position is None:
+            return
+
+        track_position = track.localize(self.car_position)
+        car_direction = track_position.faces_forward(self.car_orientation)
+        if car_direction != self.drive_forward:
+            self.steps_with_wrong_orientation += 1
+        else:
+            self.steps_with_wrong_orientation = 0
+
+        if self.steps_with_wrong_orientation > 2:
+            self.is_terminal_step = True
+
+    # Do an episode step every time laser scans are recieved
+    def on_receive_laser_scan(self, message):
+        state = self.convert_laser_message_to_tensor(message)
+
+        if self.state is not None:
+            self.check_car_orientation()
+            reward = self.get_reward()
+            self.cumulative_reward += reward
+            self.policy1.reward_episode.append(reward)
+            self.memory.append((self.state, self.action, reward, state, self.is_terminal_step))  # nopep8
+
+        if self.is_terminal_step or self.episode_length >= MAX_EPISODE_LENGTH:
+            self.drive_forward = random.random() > 0.5
+            reset_car.reset_random(
+                max_angle=math.pi / 180 * 20,
+                max_offset_from_center=0.2,
+                forward=self.drive_forward)
+            self.is_terminal_step = False
+            self.state = None
+            if self.episode_length != 0:
+                self.on_complete_episode()
+        else:
+            self.state = state
+            self.action = self.select_action(state)
+            self.perform_action(self.action)
+            self.episode_length += 1
+            self.total_step_count += 1
+
+    def on_model_state_callback(self, message):
+        if len(message.pose) < 2:
+            return
+        self.car_position = Point(
+            message.pose[1].position.x,
+            message.pose[1].position.y)
+        self.car_orientation = message.pose[1].orientation
+
+
+rospy.init_node('q_learning_training', anonymous=True)
+node = QLearningTrainingNode()
+rospy.spin()
